requirements.txt
appdirs==1.4.3
certifi==2017.4.17
future==0.16.0
nameparser==0.5.2
nltk==3.2.2
numpy==1.12.1
packaging==16.8
pyparsing==2.2.0
python-telegram-bot==5.3.0
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
urllib3==1.20
